Move the dbt templater into a plugin
