+++ plugins/sqlfluff-templater-dbt/src/dbt_templater/__init__.py
@@ -1,0 +1,10 @@
+"""Defines the hook endpoints for the dbt templater plugin."""
+
+from dbt_templater.templater import DbtTemplater
+from sqlfluff.core.plugin import hookimpl
+
+
+@hookimpl
+def get_templaters():
+    """Get templaters."""
+    return [DbtTemplater]
