+++ pytest.ini
@@ -1,3 +1,3 @@
 [pytest]
 markers =
-    dbt: marks tests needing the "dbt" extras (deselect with '-m "not dbt"')
+    dbt: marks tests needing the "dbt" plugin (deselect with '-m "not dbt"')
