+++ src/sqlfluff/config.ini
@@ -1,4 +1,4 @@
 # Config file for SQLFluff, mostly for version info for now
 [sqlfluff]
-version=0.6.2
+version=0.7.0a1
 stable_version=0.6.2
