+++ test/core/templaters/dbt_test.py
@@ -6,7 +6,6 @@ import logging
 
 from sqlfluff.core import FluffConfig, Lexer, Linter
 from sqlfluff.core.errors import SQLTemplaterSkipFile
-from sqlfluff.core.templaters.dbt import DbtTemplater
 from test.fixtures.dbt.templater import (  # noqa: F401
     DBT_FLUFF_CONFIG,
     dbt_templater,
@@ -90,11 +89,11 @@ select * from a
     ],
 )
 def test__templater_dbt_slice_file_wrapped_test(
-    raw_file, templated_file, result, caplog
+    raw_file, templated_file, result, dbt_templater, caplog
 ):
     """Test that wrapped queries are sliced safely using _check_for_wrapped()."""
     with caplog.at_level(logging.DEBUG, logger="sqlfluff.templater"):
-        _, resp, _ = DbtTemplater.slice_file(
+        _, resp, _ = dbt_templater.slice_file(
             raw_file,
             templated_file,
         )
