+++ tox.ini
@@ -14,11 +14,15 @@ deps =
     requests
     # Add the example plugin
     plugins/sqlfluff-plugin-example
+    # Install the main project
+    -e .
     # Define dbt versions
     dbt017: dbt==0.17.2
     dbt018: dbt==0.18.2
     dbt019: dbt==0.19.1
     dbt020: dbt==0.20.0
+    # Install the plugins as required
+    dbt017,dbt018,dbt019,dbt020: plugins/sqlfluff-templater-dbt
 # Include any other steps necessary for testing below.
 # {posargs} is there to allow us to specify specific tests, which
 # can then be invoked from tox by calling e.g.
